Move `StreamActions` into `StreamsOverview` directory.
